consumer/requirements.txt
confluent-kafka==2.5.0
dnspython==2.6.1
pymongo==4.8.0
pytz==2024.1
PyYAML==6.0.1
